% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{av_long_uninfected}
\alias{av_long_uninfected}
\title{Average longevity: estimate for uninfected hosts}
\usage{
av_long_uninfected(a1, b1, d1 = "")
}
\arguments{
\item{a1, b1}{numeric: location & scale parameters for background mortality,
respectively}

\item{d1}{character: probability distribution chosen to describe data}
}
\value{
a vector
}
\description{
Calculates expected average longevity due only to background mortality
}
\details{
The expected average longevity is calculated as the integral from
  zero to infinity of the cumulative survival function for background
  mortality,
  given values of a1, b1, d1
}
\examples{
  av_long_uninfected(a1 = 3.0, b1 = 0.6, d1 = "Weibull") #17.947
}
\seealso{
\code{\link{av_long_infected}}
}
